Move encoding of traces to opentelemetry-exporter-otlp-proto-common
